Move hooks out
