+++ sections/earn/Grids/StakeGrid.tsx
@@ -14,7 +14,7 @@ import {
 	SplitColumn,
 	LiquidityAmount,
 } from '../common';
-import { useRewardsTimer } from '../hooks';
+import useRewardsTimer from 'hooks/useRewardsTimer';
 
 const DEADLINE = new Date('2022-03-20T23:59:59Z');
