--- sections/earn/hooks/index.ts
@@ -1,1 +1,0 @@
-export { default as useRewardsTimer } from './useRewardsTimer';
