+++ sections/futures/MarketDetails/MarketDetails.tsx
@@ -8,13 +8,12 @@ import { wei } from '@synthetixio/wei';
 import useSelectedPriceCurrency from 'hooks/useSelectedPriceCurrency';
 import useGetFuturesMarkets from 'queries/futures/useGetFuturesMarkets';
 import useGetFuturesTradingVolume from 'queries/futures/useGetFuturesTradingVolume';
-import { FuturesMarket } from 'queries/futures/types';
 import { isFiatCurrency } from 'utils/currencies';
 import { formatCurrency, formatPercent, zeroBN } from 'utils/formatters/number';
 import useGetFuturesDailyTradeStatsForMarket from 'queries/futures/useGetFuturesDailyTrades';
 import useGetAverageFundingRateForMarket from 'queries/futures/useGetAverageFundingRateForMarket';
 import useLaggedDailyPrice from 'queries/rates/useLaggedDailyPrice';
-import { Price, Rates } from 'queries/rates/types';
+import { Rates } from 'queries/rates/types';
 import { NO_VALUE } from 'constants/placeholder';
 import StyledTooltip from 'components/Tooltip/StyledTooltip';
 import { getDisplayAsset, isEurForex, MarketKeyByAsset } from 'utils/futures';
@@ -41,11 +40,10 @@ const MarketDetails: React.FC<MarketDetailsProps> = ({ mobile }) => {
 	const futuresMarketsQuery = useGetFuturesMarkets();
 	const futuresTradingVolumeQuery = useGetFuturesTradingVolume(marketAsset);
 
-	const marketSummary: FuturesMarket | null =
-		futuresMarketsQuery?.data?.find(({ asset }) => asset === marketAsset) ?? null;
+	const marketSummary = futuresMarketsQuery.data?.find(({ asset }) => asset === marketAsset);
 
 	const futureRates = futuresMarketsQuery.isSuccess
-		? futuresMarketsQuery?.data?.reduce((acc: Rates, { asset, price }) => {
+		? futuresMarketsQuery.data?.reduce((acc: Rates, { asset, price }) => {
 				acc[MarketKeyByAsset[asset]] = price;
 				return acc;
 		  }, {})
@@ -87,11 +85,11 @@ const MarketDetails: React.FC<MarketDetailsProps> = ({ mobile }) => {
 			: undefined;
 
 	const dailyPriceChangesQuery = useLaggedDailyPrice(
-		futuresMarketsQuery?.data?.map(({ asset }) => asset) ?? []
+		futuresMarketsQuery.data?.map(({ asset }) => asset) ?? []
 	);
-	const dailyPriceChanges = dailyPriceChangesQuery?.data ?? [];
+	const dailyPriceChanges = dailyPriceChangesQuery.data ?? [];
 
-	const pastPrice = dailyPriceChanges.find((price: Price) => price.synth === marketAsset);
+	const pastPrice = dailyPriceChanges.find((price) => price.synth === marketAsset);
 
 	const assetName = `${getDisplayAsset(marketAsset)}-PERP`;
 	const fundingTitle = React.useMemo(
@@ -218,17 +216,17 @@ const MarketDetails: React.FC<MarketDetailsProps> = ({ mobile }) => {
 					marketSummary?.price && pastPrice?.price
 						? `${formatCurrency(
 								selectedPriceCurrency.name,
-								marketSummary?.price.sub(pastPrice?.price) ?? zeroBN,
+								marketSummary.price.sub(pastPrice.price) ?? zeroBN,
 								{ sign: '$', minDecimals }
 						  )} (${formatPercent(
-								marketSummary?.price.sub(pastPrice?.price).div(marketSummary?.price) ?? zeroBN
+								marketSummary.price.sub(pastPrice.price).div(marketSummary.price) ?? zeroBN
 						  )})`
 						: NO_VALUE,
 				color:
 					marketSummary?.price && pastPrice?.price
-						? marketSummary?.price.sub(pastPrice?.price).gt(zeroBN)
+						? marketSummary.price.sub(pastPrice.price).gt(zeroBN)
 							? 'green'
-							: marketSummary?.price.sub(pastPrice?.price).lt(zeroBN)
+							: marketSummary.price.sub(pastPrice.price).lt(zeroBN)
 							? 'red'
 							: ''
 						: undefined,
